fix(server): fix failing tests
