+++ server/src/utils/CustomJoi.js
@@ -22,7 +22,7 @@ const stringOperatorSchema = Joi.string().valid(...Object.keys(allOperatorsMap))
  * CustomJoi.operator(operatorSchema)
  *  * operatorSchema - The Joi-schema that the operator part of the filter should be validated against
  */
-const FilterJoi = Joi.extend({
+module.exports = Joi.extend({
     type: 'filter',
     messages: {
         'filter.base': '{{#label}} is not a valid filter',
@@ -113,15 +113,3 @@ const FilterJoi = Joi.extend({
         },
     },
 })
-
-const StringArrayJoi = Joi.extend({
-    base: Joi.array(),
-    type: 'stringArray',
-    coerce: (value, state, options) => ({
-        value: value.split ? value.split(',') : value
-    })
-})
-
-exports = module.exports = FilterJoi
-exports.FilterJoi = FilterJoi
-exports.StringArrayJoi = StringArrayJoi
